refactor: :recycle: extract functionality
