+++ client/packages/portal-client/src/components/portal-menu/AppCard/AppCard.tsx
@@ -2,6 +2,7 @@ import { Icon } from '@equinor/eds-core-react';
 import { tokens } from '@equinor/eds-tokens';
 import { menuFavoritesController } from '@equinor/portal-core';
 import { useObservable } from '@equinor/portal-utils';
+import { useNavigate } from 'react-router-dom';
 import { map } from 'rxjs';
 import styled from 'styled-components';
