+++ client/packages/portal-core/src/menu/index.ts
@@ -1,1 +1,2 @@
 export * from './menuFavorites';
+export * from './menuController';
